Initial server commit
This is the first version of our server system, it astamblishes a basic connection between the unity-client and this nodeJS server. It also already contains a little match-making system. The client will connect to the port 420 where the "proxy" is hosted, this proxy will send a redirect packet to a playable server!